journal-file: fix type of array counter

+++ src/libsystemd/sd-journal/journal-file.c
@@ -2248,7 +2248,7 @@ int journal_file_append_entry(
                 const dual_timestamp *ts,
                 const sd_id128_t *boot_id,
                 const struct iovec iovec[],
-                unsigned n_iovec,
+                size_t n_iovec,
                 uint64_t *seqnum,
                 Object **ret_object,
                 uint64_t *ret_offset) {
